+++ packages/library/src/commonMain/kotlin/io/realm/RealmList.kt
@@ -38,6 +38,7 @@ import kotlin.reflect.KClass
  * that inject values into a class. Unmanaged elements in a list can be added to a Realm using the
  * [MutableRealm.copyToRealm] method.
  */
+// FIXME Implement Versioned, RealmLifeCycle or what ever suits
 class RealmList<E> private constructor(
     delegate: MutableList<E>
 ) : MutableList<E> by delegate {
@@ -50,7 +51,7 @@ class RealmList<E> private constructor(
     /**
      * Constructs a RealmList in managed mode. For internal use only.
      */
-    constructor(
+    internal constructor(
         listPtr: NativePointer,
         metadata: OperatorMetadata
     ) : this(ManagedListDelegate(listPtr, metadata))
@@ -58,7 +59,7 @@ class RealmList<E> private constructor(
     /**
      * Metadata needed to correctly instantiate a list operator. For internal use only.
      */
-    data class OperatorMetadata(
+    internal data class OperatorMetadata(
         val clazz: KClass<*>,
         val isRealmObject: Boolean,
         val mediator: Mediator,