+++ packages/library/src/commonMain/kotlin/io/realm/internal/MutableRealmImpl.kt
@@ -28,7 +28,7 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlin.reflect.KClass
 
-class MutableRealmImpl : BaseRealmImpl, MutableRealm {
+internal class MutableRealmImpl : BaseRealmImpl, MutableRealm {
 
     // TODO Also visible as a companion method to allow for `RealmObject.delete()`, but this
     //  has drawbacks. See https://github.com/realm/realm-kotlin/issues/181