+++ packages/library/src/commonMain/kotlin/io/realm/internal/RealmObjectInternal.kt
@@ -17,6 +17,7 @@
 package io.realm.internal
 
 import io.realm.RealmObject
+import io.realm.VersionId
 import io.realm.isValid
 
 /**
@@ -27,7 +28,7 @@ import io.realm.isValid
  * [RealmObject].
  */
 @Suppress("VariableNaming")
-interface RealmObjectInternal : RealmObject, io.realm.interop.RealmObjectInterop {
+internal interface RealmObjectInternal : RealmObject, RealmLifeCycleHolder, io.realm.interop.RealmObjectInterop {
     // Names must match identifiers in compiler plugin (plugin-compiler/io.realm.compiler.Identifiers.kt)
 
     // Reference to the public Realm instance and internal transaction to which the object belongs.
@@ -35,9 +36,21 @@ interface RealmObjectInternal : RealmObject, io.realm.interop.RealmObjectInterop
     var `$realm$TableName`: String?
     var `$realm$IsManaged`: Boolean
     var `$realm$Mediator`: Mediator?
+
+    override fun realmLifeCycle(): RealmLifeCycle {
+        return `$realm$Owner` ?: UnmanagedLifeCycle
+    }
+
+    override fun version(): VersionId {
+        return super<RealmLifeCycleHolder>.version()
+    }
 }
 
-fun RealmObjectInternal.checkValid() {
+internal inline fun RealmObject.realmObjectInternal(): RealmObjectInternal {
+    return this as RealmObjectInternal
+}
+
+internal fun RealmObjectInternal.checkValid() {
     if (!this.isValid()) {
         throw IllegalStateException("Cannot perform this operation on an invalid/deleted object")
     }